--- game/Assets/Plugins/Sentry/SentryCliOptions.asset
@@ -1,19 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 582302131, guid: 7f7ea26e3243148e5ab9a7c58391af93, type: 3}
-  m_Name: SentryCliOptions
-  m_EditorClassIdentifier: 
-  <UploadSymbols>k__BackingField: 1
-  <UploadDevelopmentSymbols>k__BackingField: 0
-  <Auth>k__BackingField: 
-  <Organization>k__BackingField: 
-  <Project>k__BackingField:
