--- game/Assets/Samples/Sentry Unity/0.7.0/Unity of bugs/Assets/Scenes/1_BugfarmSettings.lighting.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5f9e4748960584c51bee9b1d3f39fee6
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 4890085278179872738
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
