--- game/Assets/Samples/Sentry Unity/0.7.0/Unity of bugs/Assets/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b276c8a017c1344969750250cc397936
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
