+++ game/Assets/Samples/Sentry Unity/0.8.0.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b1c082e8477514bfa84aec1f167eacb7
+guid: 34cd06e3ee59c439e8458eeff7869eeb
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
