+++ game/Assets/Samples/Sentry Unity/0.8.0/Unity of bugs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b8afd42f9ea45460283d5209e6653024
+guid: 152292976a5c14632b866878186038a1
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
