+++ game/Assets/Samples/Sentry Unity/0.8.0/Unity of bugs/Scenes.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e234c71f4b79e481d8d7a774456651f5
+guid: 67079b374a31648218ba960c327c7139
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
