+++ game/Assets/Samples/Sentry Unity/0.8.0/Unity of bugs/Scripts.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fc0ee7b5d1cd5446cbb5fc284a5f1a16
+guid: 1aa08bd4bf4d64dbea0fa07b242312bd
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
