+++ game/Assets/Samples/Sentry Unity/0.8.0/Unity of bugs/Scripts/MobileNativeSupport/CppPlugin.cpp.meta
